Support for static file overrides
Some applications can't handle moving them to a different location
and sometimes small patching of javascript assets (and similar)
will have to be done. While that should be done at the source, it's
sometimes difficult since the assets are well embedded within a
war file (that is difficult to modify). This offers an alternative
to hijack a file and serve a local copy instead.

Originally done to allow Zeppelin 0.5.0 to be served from a different
directory than / and let it connect to a websocket path that is
not /ws.