Improved authentication (server auth)
When the user is not authenticated, which we decide based on a
cookie that we set, we redirect the user to an authentication
server. That server will make the user sign in and redirect him
back to our server, together with a code.

We then use that code + host + path to ask the authentication
server if this is a valid request.